towerpanel-backend/requirements.txt
fastapi
pyserial-asyncio
uvicorn[standard]
